refactor: reservation controller 로직 일부 service 내부로 이동
