+++ src/main/java/roomescape/presentation/web/controller/ReservationController.java
@@ -4,8 +4,6 @@ import java.net.URI;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
-import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -13,10 +11,10 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.domain.Name;
+import roomescape.application.ReservationService;
+import roomescape.application.request.CreateReservationRequest;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
-import roomescape.presentation.web.request.CreateReservationRequest;
 import roomescape.presentation.web.response.ReservationResponse;
 import roomescape.presentation.web.response.ReservationTimeResponse;
 
@@ -24,81 +22,38 @@ import roomescape.presentation.web.response.ReservationTimeResponse;
 @RequestMapping("/reservations")
 class ReservationController {
 
-    private final JdbcTemplate jdbcTemplate;
-    private final SimpleJdbcInsert simpleInsert;
+    private final ReservationService reservationService;
 
-    public ReservationController(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-        this.simpleInsert = new SimpleJdbcInsert(jdbcTemplate.getDataSource())
-                .withTableName("reservation")
-                .usingGeneratedKeyColumns("id");
+    public ReservationController(JdbcTemplate jdbcTemplate, ReservationService reservationService) {
+        this.reservationService = reservationService;
     }
 
     @PostMapping
     public ResponseEntity<ReservationResponse> reserve(@RequestBody CreateReservationRequest request) {
-        BeanPropertySqlParameterSource parameterSource = new BeanPropertySqlParameterSource(request);
-        Number key = simpleInsert.executeAndReturnKey(parameterSource);
-
         // create new Reservation
-        Reservation newReservation = new Reservation(
-                key.longValue(),
-                new Name(request.name()),
-                request.date(),
-                resolveTime(request.timeId())
-        );
+        Reservation newReservation = reservationService.reserve(request);
 
-        // build response dto
+        // build web response dto
         ReservationResponse response = toDto(newReservation);
-        return ResponseEntity.created(URI.create("/reservations/" + key))
+        return ResponseEntity.created(URI.create("/reservations/" + newReservation.getId()))
                 .body(response);
-    }
-
-    private ReservationTime resolveTime(Long timeId) {
-        String sql = "select * from reservation_time where id = ?";
-        ReservationTime reservationTime = jdbcTemplate.queryForObject(sql, (rs, rowNum) -> new ReservationTime(
-                rs.getLong("id"),
-                rs.getTime("start_at").toLocalTime()
-        ), timeId);
-
-        return reservationTime;
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteBy(@PathVariable Long id) {
-        String sql = "delete from reservation where id = ?";
-        jdbcTemplate.update(sql, id);
+        reservationService.deleteBy(id);
         return ResponseEntity.noContent().build();
     }
 
     @GetMapping
     public ResponseEntity<List<ReservationResponse>> getReservations() {
-        String sql = """
-                    select 
-                        r.id as reservation_id,
-                        r.name as reservation_name,
-                        r.date as reservation_date,
-                        t.id as time_id,
-                        t.start_at as time_value
-                    from reservation as r
-                    inner join reservation_time as t
-                    on r.time_id = t.id
-                """;
+        List<Reservation> reservations = reservationService.findReservations();
 
-        List<Reservation> reservations = jdbcTemplate.query(sql, (rs, rowNum) ->
-                new Reservation(
-                        rs.getLong("reservation_id"),
-                        new Name(rs.getString("reservation_name")),
-                        rs.getDate("reservation_date").toLocalDate(),
-                        new ReservationTime(
-                                rs.getLong("time_id"),
-                                rs.getTime("time_value").toLocalTime()
-                        )
-                )
-        );
-
-        return ResponseEntity.ok(reservations.stream()
+        // build web response dto
+        List<ReservationResponse> responses = reservations.stream()
                 .map(this::toDto)
-                .toList());
+                .toList();
+        return ResponseEntity.ok(responses);
     }
 
     private ReservationResponse toDto(Reservation newReservation) {
